chore: remove plan upload metadata operation
