+++ src/plugins/bundlrStorage/index.ts
@@ -1,3 +1,2 @@
 export * from './BundlrStorageDriver';
-export * from './planUploadMetadataUsingBundlrOperationHandler';
 export * from './plugin';
