+++ src/plugins/nftModule/uploadMetadata.ts
@@ -1,8 +1,9 @@
+import cloneDeep from 'lodash.clonedeep';
 import { Metaplex } from '@/Metaplex';
 import { Operation, OperationHandler, useOperation } from '@/types';
+import { walk } from '@/utils';
 import { JsonMetadata } from './JsonMetadata';
-import { planUploadMetadataOperation } from './planUploadMetadata';
-import { MetaplexFile } from '../storageModule';
+import { isMetaplexFile, MetaplexFile } from '../storageModule';
 
 const Key = 'UploadMetadataOperation' as const;
 export const uploadMetadataOperation =
@@ -17,6 +18,7 @@ export type UploadMetadataInput = JsonMetadata<MetaplexFile | string>;
 
 export interface UploadMetadataOutput {
   metadata: JsonMetadata;
+  assetUris: string[];
   uri: string;
 }
 
@@ -26,10 +28,46 @@ export const uploadMetadataOperationHandler: OperationHandler<UploadMetadataOper
       operation: UploadMetadataOperation,
       metaplex: Metaplex
     ): Promise<UploadMetadataOutput> => {
-      const plan = await metaplex
-        .operations()
-        .execute(planUploadMetadataOperation(operation.input));
+      const rawMetadata = operation.input;
+      const files = getAssetsFromJsonMetadata(rawMetadata);
+      const assetUris = await metaplex.storage().uploadAll(files);
+      const metadata = replaceAssetsWithUris(rawMetadata, assetUris);
+      const uri = await metaplex.storage().uploadJson(metadata);
 
-      return plan.execute();
+      return { uri, metadata, assetUris };
     },
   };
+
+export const getAssetsFromJsonMetadata = (
+  input: UploadMetadataInput
+): MetaplexFile[] => {
+  const files: MetaplexFile[] = [];
+
+  walk(input, (next, value) => {
+    if (isMetaplexFile(value)) {
+      files.push(value);
+    } else {
+      next(value);
+    }
+  });
+
+  return files;
+};
+
+export const replaceAssetsWithUris = (
+  input: UploadMetadataInput,
+  replacements: string[]
+): JsonMetadata => {
+  const clone = cloneDeep(input);
+  let index = 0;
+
+  walk(clone, (next, value, key, parent) => {
+    if (isMetaplexFile(value) && index < replacements.length) {
+      parent[key] = replacements[index++];
+    }
+
+    next(value);
+  });
+
+  return clone as JsonMetadata;
+};
